Extract internal-transfer adjustment helper from SnapshotService
Move internal_transfer_adjustments_by_date_base out of SnapshotService
into a standalone portfolio::transfers module. The function is logically
about activities + fx, not snapshots, and will need to be called from
the portfolio valuation aggregator. Pure refactor: the existing
SnapshotService method now delegates to the helper and behavior is
unchanged.
