+++ crates/core/src/portfolio/mod.rs
@@ -5,4 +5,5 @@ pub mod income;
 pub mod net_worth;
 pub mod performance;
 pub mod snapshot;
+pub mod transfers;
 pub mod valuation;
